rename gitlab/mr to merge_request
